// lib/Q/sys/Corelib.d.ts
export declare namespace string {
    function IsPrintable(keyCode: number, charCode: number): boolean;
}
export declare namespace helper {
    function TryCatch<T>(owner: any, Try: (...args: any[]) => T, Catch?: (e: Error, ...args: any[]) => T, params?: any[]): T;
}
export declare namespace basic {
// ...
    interface ICrypto {
        Encrypt(data: Uint8Array | number[]): (Uint8Array | number[]);
        Decrypt(data: Uint8Array | number[]): (Uint8Array | number[]);
        SEncrypt(data: string): string;
        SDecrypt(data: string): string;
    }
// ...
    interface IEventHandler extends IDisposable {
        Started: boolean;
        Start(): any;
        Pause(): any;
        Dispose(): any;
        Reset(): any;
    }
    interface Module {
    }
    interface IContext {
        CanAccessToMe(type: string, folder: string, name: string): any;
        GetPath(path: string): string;
        NameOf(type: Function): string;
        GetType(path: string): Function;
    }
    interface IDisposable {
        Dispose(force?: boolean): any;
    }
    interface IBindable {
        Owner?: any;
        Invoke(...args: any[]): any;
    }
    interface ITBindable<T extends (...args: any[]) => void> extends IBindable {
        Invoke: T;
    }
// ...
    interface IJob {
        Name: string;
        Todo?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        Check?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        OnError?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        OnInitialize?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        OnScopDisposing?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
    }
// ...
    interface IId {
        Id: number;
    }
// ...
}
// ...
    type Method<RET, T extends (...args: any[]) => RET> = MethodGroup<RET, T> | T | basic.ITBindable<T>;
// ...
export declare namespace bind {
    class Job implements basic.IJob {
        Name: string;
        Todo: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        Check: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        OnError: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        OnInitialize: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        OnScopDisposing: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        constructor(Name: string, Todo: (job: JobInstance, e: bind.EventArgs<any, any>) => void, Check?: (job: JobInstance, e: bind.EventArgs<any, any>) => void, OnError?: (job: JobInstance, e: bind.EventArgs<any, any>) => void, OnInitialize?: (job: JobInstance, e: bind.EventArgs<any, any>) => void, OnScopDisposing?: (job: JobInstance, e: bind.EventArgs<any, any>) => void);
    }
    class Jobs implements basic.IJob {
        Name: string;
        Todo(job: JobInstance, e: bind.EventArgs<any, any>): void;
        Check(job: JobInstance, e: bind.EventArgs<any, any>): void;
        OnError(job: JobInstance, e: bind.EventArgs<any, any>): void;
        OnInitialize(job: JobInstance, e: bind.EventArgs<any, any>): void;
        OnScopDisposing(job: JobInstance, e: bind.EventArgs<any, any>): void;
        constructor(Name: string);
        push(jobName: any): void;
    }
    class JobInstance implements EventListenerObject {
        Scop: bind.Scop;
        job: basic.IJob;
        dom: Node;
        Control: UI.JControl;
        private _events;
        Handle: (ji: JobInstance, e: Event) => void;
        addEventListener(name: string, event: string, delegate: EventListenerOrEventListenerObject | any): void;
        removeEventListener(name: string): void;
        getEvent(name: string): EventListenerOrEventListenerObject;
        constructor(Scop: bind.Scop, job: basic.IJob, dom: Node);
        private propb;
        private ValueChanged(s, e);
        Dispose(): void;
        IsDisposed: any;
        _store: any;
        addValue(name: string, value: any): void;
        getValue(name: string): any;
        Checker: (value: any) => boolean;
        Ischanging: boolean;
        handleEvent(e: Event): void;
    }
    function GetJob(name: string): basic.IJob;
    function Register(job: basic.IJob, override?: boolean): basic.IJob;
}
// ...
export declare namespace bind {
// ...
    class EventListener<T extends Function> implements basic.IDisposable {
        private _deleagtes;
        private key;
        private isSingliton;
        constructor(key: any, isSingliton?: boolean);
        On: T;
        private locks;
        Off: T;
        private lock;
        Invoke(key: Object, params: any[]): void;
        Invok(key: Object, callBack: (delegate: T) => any): any;
        PInvok(key: Object, params: any[], owner?: any): any;
        Add(delegate: T, key?: any): void;
        Remove(key: any): void;
        private _store;
        Dispose(): void;
    }
// ...
}
// ...
    interface Node<T> {
        Depth: number;
        Value: T;
        param?: any;
        children: Node<T>[];
        Parent: Node<T>;
    }
// ...
export declare namespace Api {
    interface IApiTrigger {
        Name: string;
        Filter: (cmdCallback: IApiCallback, params: any) => boolean;
        CheckAccess: (t: IApiTrigger) => boolean;
        Params?: any;
    }
    interface IApiCallback {
        hash?: string;
        Name: string;
        DoApiCallback: (trigger: IApiTrigger, callback: IApiCallback, params: IApiParam) => void;
        Owner?: any;
        Params?: any;
    }
    function RegisterApiCallback(api: IApiCallback): boolean;
    function RegisterTrigger(api: IApiTrigger): boolean;
    function RiseApi(apiName: string, params: IApiParam): void;
    interface IApiParam {
        data: any;
        callback?(p: IApiParam, args: any): any;
    }
    interface IApi {
        Trigger: IApiTrigger;
        Callback: IApiCallback[];
    }
    interface apiGarbage {
        [name: string]: IApi;
    }
}
// ...
export declare namespace net {
    class Header {
        private _key;
        readonly key: string;
        private _value;
        readonly value: string;
        constructor(key: any, value: any);
    }
    enum ResponseType {
        json = 0,
        document = 1,
        text = 2,
        arraybuffer = 3,
        blob = 4,
    }
    enum WebRequestMethod {
        Get = 0,
        Post = 1,
        Head = 2,
        Put = 3,
        Delete = 4,
        Options = 5,
        Connect = 6,
        Create = 7,
        Open = 8,
        Close = 9,
        Validate = 10,
        FastValidate = 11,
        Print = 12,
        UPDATE = 13,
        SUPDATE = 14,
        Set = 15,
    }
    class WebRequest implements basic.IDisposable {
        crypt: basic.ICrypto;
        Uid: string;
        Pwd: string;
        private http;
        private _responseType;
        getResponseType(): ResponseType;
        setResponseType(v: ResponseType): ResponseType;
        Crypto: basic.ICrypto;
        private key;
        private downloadDelegate;
        constructor(crypt?: basic.ICrypto);
        Dispose(): void;
        private _onprogress(e);
        readonly IsSuccess: boolean;
        Download(req: IRequestUrl, data: any): void;
        Download2(c: Request): void;
        private getUrlOf(c);
        private getDataOf(c);
        GetFileSize(url: any, callback: any): void;
        RequestHeader(url: any, callback: any): void;
        OnComplete: bind.EventListener<(e: WebRequest) => void>;
        readonly Response: string;
        GetHeader(name: string): string;
        GetHeaders(): string;
    }
    abstract class RequestParams<T, S> {
        protected callback: (sender: S, result: any) => void;
        data: T;
        isPrivate: boolean;
        IsSuccess: boolean;
        constructor(callback: (sender: S, result: any) => void, data: T, isPrivate?: boolean);
        Callback(sender: S, result: any): void;
        abstract OutputData(): string;
        InputData: string;
    }
    interface RequestMethodShema {
        Method: WebRequestMethod;
        Name: string;
        SendData: boolean;
        ParamsFormat?: basic.StringCompile;
    }
    interface IRequestParams {
        [name: string]: string | number | boolean;
    }
    class Request {
        url: IRequestUrl;
        data: RequestParams<any, QueeDownloader>;
        params: IRequestParams;
        fail: boolean;
        constructor(url: IRequestUrl, data: RequestParams<any, QueeDownloader>, params: IRequestParams);
    }
    class QueeDownloader {
        crypt: basic.ICrypto;
        private webr;
        Uid: string;
        Pwd: string;
        readonly Request: net.WebRequest;
        private quee;
        private isRunning;
        private isDownloading;
        Crypto: basic.ICrypto;
        constructor(crypt: basic.ICrypto);
        current: Request;
        private OnError();
        private DownloadComplete(xmlRequest);
        Push(url: IRequestUrl, data: RequestParams<any, QueeDownloader>, params: IRequestParams): void;
        Insert(dcall: Request): void;
        Start(): void;
        Next(): void;
        Restart(): void;
        OnSuccess: bind.EventListener<any>;
        OnFail: bind.EventListener<any>;
        OnFinish: bind.EventListener<any>;
    }
}
export declare namespace net {
    interface IRequestHeader {
        [key: string]: string;
    }
    interface IRequestUrl {
        beforRequest?: (req: net.IRequestUrl) => void;
        Url: string;
        Method?: net.WebRequestMethod;
        Header?: IRequestHeader;
        HasBody?: boolean;
        timeout?: number;
        ResponseType?: ResponseType;
    }
    class RequestUrl implements IRequestUrl {
        private _url;
        private context;
        Header: IRequestHeader;
        Method: net.WebRequestMethod;
        HasBody: boolean;
        ResponseType: ResponseType;
        beforRequest: (req: net.IRequestUrl) => void;
        timeout?: number;
        Url: string;
        constructor(_url: string, context: basic.IContext, Header?: IRequestHeader, Method?: net.WebRequestMethod, HasBody?: boolean, ResponseType?: ResponseType);
    }
}
export declare namespace basic {
    class DomEventHandler<T extends Event, P> implements IEventHandler, EventListenerObject {
        dom: Element;
        event: string;
        private owner;
        private handle;
        private param;
        Started: boolean;
        constructor(dom: Element, event: string, owner: any, handle: (eh: DomEventHandler<T, P>, ev: T, param: P) => void, param?: P);
        Start(): void;
        Pause(): void;
        Dispose(): void;
        Reset(): void;
        handleEvent(evt: Event): void;
        static Dispose(dom: EventTarget, event?: string): void;
    }
}